% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cal_sheet_names}
\alias{cal_sheet_names}
\title{Get sheet names from an Excel file}
\usage{
cal_sheet_names(path)
}
\arguments{
\item{path}{Path to the Excel file}
}
\value{
Character vector of sheet names
}
\description{
Get sheet names from an Excel file
}
