% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{calamineR-package}
\alias{calamineR}
\alias{calamineR-package}
\title{calamineR: Fast Excel File Reader Using Rust}
\description{
High-performance Excel file reader supporting xlsx, xlsm, xlsb, xls, and ods formats. Built on the Rust 'calamine' library, providing significantly faster read times compared to pure R implementations. Supports reading sheet names, dimensions, and data as data frames. Can optionally fill merged cells with the top-left cell value.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MaxManek1110/calamineR}
  \item Report bugs at \url{https://github.com/MaxManek1110/calamineR/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Manek \email{maximilian.manek@hotmail.de}

Authors:
\itemize{
  \item Juan Antonio Monroy Kuhn \email{amonroykuhn@gmail.com}
  \item Johann Tuffe (calamine Rust library)
  \item Authors of the dependent Rust crates (see inst/AUTHORS file)
}

}
\keyword{internal}
